Fix error when destructing a static std::unique_ptr<cppflow::model>
When creating a static version of a reference to a `cppflow::model`,
there becomes an error when destructing, due to the
`context::get_status()` method using a `thread_local` status. While
this is done normally for performance gains and is fine in most
places. However, when the program is exiting, it first destructs `thread_local`
objects and then `static` ones. This leads to `local_tf_status`
containing a `nullptr` instead of a pointer to a `TF_Status`
object. Therefore, `TF_DeleteSession(sess, status)` fails because it
access `*status`. The simplest fix is to not re-use this status object
in the destructor of the session and instead make a new,
non-`thread_local` status instead.